requirements.txt
alembic==1.0.10
Click==7.0
dominate==2.3.5
Flask==1.0.3
Flask-Bootstrap4==4.0.2
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.11
MarkupSafe==1.1.1
psycopg2-binary==2.8.2
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.4
visitor==0.1.3
Werkzeug==0.15.4
WTForms==2.2.1
